backend/requirements.txt
# AI Personal Finance Tracker - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
starlette==0.36.3

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
python-multipart==0.0.9

# Data Validation
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# File Parsing
pdfplumber==0.10.4
openpyxl==3.1.2
pandas==2.2.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# ML/AI (optional - for advanced categorization)
scikit-learn==1.4.0
numpy==1.26.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.15
mypy==1.8.0
pre-commit==3.6.0

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.1
tenacity==8.2.3
